Split agent's HTTP server and handler
