+++ cmd/agent/app/httpserver/handler.go
@@ -34,23 +34,6 @@ var (
 	errBadRequest = errors.New("bad request")
 )
 
-// NewHTTPServer creates a new server that hosts an HTTP/JSON endpoint for clients
-// to query for sampling strategies and baggage restrictions.
-func NewHTTPServer(hostPort string, manager configmanager.ClientConfigManager, mFactory metrics.Factory) *http.Server {
-	handler := newHTTPHandler(manager, mFactory)
-	mux := http.NewServeMux()
-	mux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		handler.serveSamplingHTTP(w, r, true /* thriftEnums092 */)
-	})
-	mux.HandleFunc("/sampling", func(w http.ResponseWriter, r *http.Request) {
-		handler.serveSamplingHTTP(w, r, false /* thriftEnums092 */)
-	})
-	mux.HandleFunc("/baggageRestrictions", func(w http.ResponseWriter, r *http.Request) {
-		handler.serveBaggageHTTP(w, r)
-	})
-	return &http.Server{Addr: hostPort, Handler: mux}
-}
-
 func newHTTPHandler(manager configmanager.ClientConfigManager, mFactory metrics.Factory) *httpHandler {
 	handler := &httpHandler{manager: manager}
 	metrics.Init(&handler.metrics, mFactory, nil)
